split remaining spec.c/.h into three modules
